+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 Minor changes:
  - Add footer
  - Add terms of use and privacy policy page
+ - Fix bug where user cannot access single question page
+ - Add explore users page
 
 Bug Fixes:
  - Fix CORS issue